--- apps/main/src/app/not-found.tsx
@@ -1,13 +1,0 @@
-import { t } from '@ui-kit/lib/i18n'
-import { PageNotFound } from '@ui-kit/pages/PageNotFound'
-
-export default function NotFound() {
-  return (
-    <>
-      <head>
-        <title>{t`Error 404` + ' - Curve'}</title>
-      </head>
-      <PageNotFound />
-    </>
-  )
-}
