+++ apps/main/src/routes/GlobalLayout.tsx
@@ -1,17 +1,17 @@
 import { type ReactNode } from 'react'
-import daoNetworks from '@/dao/networks'
-import { useDexAppStats, useDexRoutes } from '@/dex/hooks/useDexAppStats'
-import { useLendAppStats } from '@/lend/hooks/useLendAppStats'
-import lendNetworks from '@/lend/networks'
-import { useLlamalendAppStats } from '@/llamalend/hooks/useLendAppStats'
-import { useLoanAppStats } from '@/loan/hooks/useLoanAppStats'
-import crvusdNetworks from '@/loan/networks'
+import daoNetworks from '@/dao/networks.ts'
+import { useDexAppStats, useDexRoutes } from '@/dex/hooks/useDexAppStats.ts'
+import { useLendAppStats } from '@/lend/hooks/useLendAppStats.tsx'
+import lendNetworks from '@/lend/networks.ts'
+import { useLlamalendAppStats } from '@/llamalend/hooks/useLendAppStats.tsx'
+import { useLoanAppStats } from '@/loan/hooks/useLoanAppStats.tsx'
+import crvusdNetworks from '@/loan/networks.ts'
 import type { IChainId as LlamaChainId } from '@curvefi/llamalend-api/lib/interfaces'
 import Box from '@mui/material/Box'
 import Stack from '@mui/material/Stack'
 import type { NetworkDef, NetworkMapping } from '@ui/utils'
-import { APP_LINK, AppMenuOption, type AppName } from '@ui-kit/shared/routes'
-import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
+import { APP_LINK, AppMenuOption, type AppName } from '@ui-kit/shared/routes.ts'
+import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces.ts'
 import { Footer } from '@ui-kit/widgets/Footer'
 import { Header as Header } from '@ui-kit/widgets/Header'
