+++ apps/main/src/routes/index.tsx
@@ -1,17 +1,23 @@
-import { lazy } from 'react'
 import { rootRoute } from '@/routes/root.routes'
 import { redirectTo } from '@/routes/util'
+import Skeleton from '@mui/material/Skeleton'
 import { createRoute, createRouter } from '@tanstack/react-router'
+import { t } from '@ui-kit/lib/i18n.ts'
+import { PageNotFound } from '@ui-kit/pages/PageNotFound.tsx'
+import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces.ts'
 import { crvusdRoutes } from './crvusd.routes'
 import { daoRoutes } from './dao.routes'
 import { dexRoutes } from './dex.routes'
 import { lendRoutes } from './lend.routes'
 import { llamalendRoutes } from './llamalend.routes'
 
+const { MinHeight } = SizesAndSpaces
+
 const indexRoute = createRoute({
   getParentRoute: () => rootRoute,
   path: '/',
-  component: lazy(() => import('../app/page')),
+  /** Redirect is handled by the `RootLayout` component */
+  component: () => <Skeleton width="100%" height={MinHeight.pageContent} />,
   head: () => ({
     meta: [{ title: 'Curve.finance' }],
   }),
@@ -35,7 +41,14 @@ export const router = createRouter({
   ]),
   defaultPreload: 'intent',
   defaultPreloadStaleTime: 0,
-  defaultNotFoundComponent: lazy(() => import('../app/not-found')),
+  defaultNotFoundComponent: () => (
+    <>
+      <head>
+        <title>{t`Error 404` + ' - Curve'}</title>
+      </head>
+      <PageNotFound />
+    </>
+  ),
 })
 
 // Register router for type safety
